+++ src/main/java/beam/physsim/jdeqsim/cacc/jdeqsim/EnterRoadMessage.java
@@ -17,26 +17,44 @@
  *                                                                         *
  * *********************************************************************** */
 
-package beam.playground.jdeqsim_with_cacc.jdeqsim;
+package beam.physsim.jdeqsim.cacc.jdeqsim;
+
+import org.matsim.api.core.v01.Id;
+import org.matsim.api.core.v01.events.Event;
+import org.matsim.api.core.v01.events.LinkEnterEvent;
+import org.matsim.api.core.v01.events.VehicleEntersTrafficEvent;
 
 /**
- * The basic EventMessage type.
+ * The micro-simulation internal handler for entering a road.
  *
  * @author rashid_waraich
  */
-public abstract class EventMessage extends Message {
-	public Vehicle vehicle;
-	public Scheduler scheduler;
+public class EnterRoadMessage extends EventMessage {
 
-	public EventMessage(Scheduler scheduler, Vehicle vehicle) {
-		super();
-		this.vehicle = vehicle;
-		this.scheduler = scheduler;
+	@Override
+	public void handleMessage() {
+		// enter the next road
+		Road road = Road.getRoad(vehicle.getCurrentLinkId());
+		road.enterRoad(vehicle, getMessageArrivalTime());
 	}
 
-	public void resetMessage(Scheduler scheduler, Vehicle vehicle) {
-		this.scheduler = scheduler;
-		this.vehicle = vehicle;
+	public EnterRoadMessage(Scheduler scheduler, Vehicle vehicle) {
+		super(scheduler, vehicle);
+		priority = JDEQSimConfigGroup.PRIORITY_ENTER_ROAD_MESSAGE;
+	}
+
+	@Override
+	public void processEvent() {
+		Event event = null;
+
+		// the first EnterLink in a leg is a Wait2LinkEvent
+		if (vehicle.getLinkIndex() == -1) {
+			event = new VehicleEntersTrafficEvent(this.getMessageArrivalTime(), vehicle.getOwnerPerson().getId(), vehicle.getCurrentLinkId(), 
+					Id.create(vehicle.getOwnerPerson().getId(), org.matsim.vehicles.Vehicle.class), null, 1.0);
+		} else {
+			event = new LinkEnterEvent(this.getMessageArrivalTime(), Id.create(vehicle.getOwnerPerson().getId(), org.matsim.vehicles.Vehicle.class), vehicle.getCurrentLinkId());
+		}
+		eventsManager.processEvent(event);
 	}
 
 }
