+++ src/main/java/beam/physsim/jdeqsim/cacc/jdeqsim/Message.java
@@ -17,7 +17,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package beam.playground.jdeqsim_with_cacc.jdeqsim;
+package beam.physsim.jdeqsim.cacc.jdeqsim;
 
 import org.matsim.core.api.experimental.events.EventsManager;
